requirements.txt
streamlit==1.32.0
pymongo==4.6.1
pandas==2.2.1
numpy==1.26.4
biopython==1.83
scikit-learn==1.4.0
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
plotly==5.19.0
requests==2.31.0
py3Dmol==2.0.4
streamlit-autorefresh==1.0.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2 